requirements.txt
discord.py
aiohttp
requests
.env
